+++ pipelex/core/pipes/pipe_blueprint.py
@@ -22,7 +22,7 @@ class AllowedPipeTypes(StrEnum):
     # Pipe Operators
     PIPE_FUNC = "PipeFunc"
     PIPE_IMG_GEN = "PipeImgGen"
-    PIPE_JINJA2 = "PipeJinja2"
+    PIPE_COMPOSE = "PipeCompose"
     PIPE_LLM = "PipeLLM"
     PIPE_OCR = "PipeOcr"
     # Pipe Controller