+++ pipelex/migration/migrate_v0_1_0_to_v0_2_0.py
@@ -18,7 +18,7 @@ class TOMLMigrator:
         "PipeOcr",
         "PipeImgGen",
         "PipeFunc",
-        "PipeJinja2",
+        "PipeCompose",
         "PipeLLMPrompt",
         "PipeSequence",
         "PipeBatch",