+++ pipelex/pipe_operators/compose/pipe_compose_blueprint.py
@@ -4,6 +4,6 @@ from pipelex.core.pipes.pipe_blueprint import PipeBlueprint
 from pipelex.tools.templating.jinja2_blueprint import Jinja2Blueprint
 
 
-class PipeJinja2Blueprint(PipeBlueprint, Jinja2Blueprint):
-    type: Literal["PipeJinja2"] = "PipeJinja2"
+class PipeComposeBlueprint(PipeBlueprint, Jinja2Blueprint):
+    type: Literal["PipeCompose"] = "PipeCompose"
     category: Literal["PipeOperator"] = "PipeOperator"