+++ pipelex/pipe_operators/compose/pipe_compose_factory.py
@@ -6,23 +6,23 @@ from pipelex.core.pipes.pipe_input import PipeInputSpec
 from pipelex.core.pipes.pipe_input_factory import PipeInputSpecFactory
 from pipelex.exceptions import PipeDefinitionError
 from pipelex.hub import get_concept_provider
-from pipelex.pipe_operators.jinja2.pipe_jinja2 import PipeJinja2
-from pipelex.pipe_operators.jinja2.pipe_jinja2_blueprint import PipeJinja2Blueprint
+from pipelex.pipe_operators.compose.pipe_compose import PipeCompose
+from pipelex.pipe_operators.compose.pipe_compose_blueprint import PipeComposeBlueprint
 from pipelex.tools.templating.jinja2_parsing import check_jinja2_parsing
 from pipelex.tools.templating.jinja2_template_category import Jinja2TemplateCategory
 from pipelex.tools.templating.template_preprocessor import preprocess_template
 
 
-class PipeJinja2Factory(PipeFactoryProtocol[PipeJinja2Blueprint, PipeJinja2]):
+class PipeComposeFactory(PipeFactoryProtocol[PipeComposeBlueprint, PipeCompose]):
     @classmethod
     @override
     def make_from_blueprint(
         cls,
         domain: str,
         pipe_code: str,
-        blueprint: PipeJinja2Blueprint,
+        blueprint: PipeComposeBlueprint,
         concept_codes_from_the_same_domain: list[str] | None = None,
-    ) -> PipeJinja2:
+    ) -> PipeCompose:
         preprocessed_template: str | None = None
         if blueprint.jinja2:
             preprocessed_template = preprocess_template(blueprint.jinja2)
@@ -38,7 +38,7 @@ class PipeJinja2Factory(PipeFactoryProtocol[PipeJinja2Blueprint, PipeJinja2]):
             concept_string_or_code=blueprint.output,
             concept_codes_from_the_same_domain=concept_codes_from_the_same_domain,
         )
-        return PipeJinja2(
+        return PipeCompose(
             domain=domain,
             code=pipe_code,
             definition=blueprint.definition,
@@ -61,31 +61,32 @@ class PipeJinja2Factory(PipeFactoryProtocol[PipeJinja2Blueprint, PipeJinja2]):
         )
 
     @classmethod
-    def make_pipe_jinja2_from_template_str(
+    def make_pipe_compose_from_template_str(
         cls,
         domain: str,
         inputs: PipeInputSpec | None = None,
         template_str: str | None = None,
         template_name: str | None = None,
-    ) -> PipeJinja2:
+    ) -> PipeCompose:
         if template_str:
             preprocessed_template = preprocess_template(template_str)
             check_jinja2_parsing(
                 jinja2_template_source=preprocessed_template,
                 template_category=Jinja2TemplateCategory.LLM_PROMPT,
             )
-            return PipeJinja2(
+            return PipeCompose(
                 domain=domain,
-                code="adhoc_pipe_jinja2_from_template_str",
+                code="adhoc_pipe_compose_from_template_str",
                 jinja2=preprocessed_template,
                 inputs=inputs or PipeInputSpecFactory.make_empty(),
             )
-        if template_name:
-            return PipeJinja2(
+        elif template_name:
+            return PipeCompose(
                 domain=domain,
-                code="adhoc_pipe_jinja2_from_template_name",
+                code="adhoc_pipe_compose_from_template_name",
                 jinja2_name=template_name,
                 inputs=inputs or PipeInputSpecFactory.make_empty(),
             )
-        msg = "Either template_str or template_name must be provided to make_pipe_jinja2_from_template_str"
-        raise PipeDefinitionError(msg)
+        else:
+            msg = "Could not make a PipeCompose because neither template_str nor template_name were provided"
+            raise PipeDefinitionError(msg)