+++ tests/unit/pipelex/core/test_data/pipes/operators/compose/pipe_compose.py
@@ -1,15 +1,15 @@
 from pipelex.core.bundles.pipelex_bundle_blueprint import PipelexBundleBlueprint
 from pipelex.core.concepts.concept_native import NativeConceptEnum
-from pipelex.pipe_operators.jinja2.pipe_jinja2_blueprint import PipeJinja2Blueprint
+from pipelex.pipe_operators.compose.pipe_compose_blueprint import PipeComposeBlueprint
 from pipelex.tools.templating.jinja2_template_category import Jinja2TemplateCategory
 
-PIPE_JINJA2 = (
-    "pipe_jinja2",
+PIPE_COMPOSE = (
+    "pipe_compose",
     """domain = "test_pipes"
 definition = "Domain with template processing pipe"
 
 [pipe.process_template]
-type = "PipeJinja2"
+type = "PipeCompose"
 definition = "Process a Jinja2 template"
 output = "Text"
 jinja2 = "Hello {{ name }}!"
@@ -19,8 +19,8 @@ template_category = "markdown"
         domain="test_pipes",
         definition="Domain with template processing pipe",
         pipe={
-            "process_template": PipeJinja2Blueprint(
-                type="PipeJinja2",
+            "process_template": PipeComposeBlueprint(
+                type="PipeCompose",
                 definition="Process a Jinja2 template",
                 output=NativeConceptEnum.TEXT,
                 jinja2="Hello {{ name }}!",
@@ -30,7 +30,7 @@ template_category = "markdown"
     ),
 )
 
-# Export all PipeJinja2 test cases
-PIPE_JINJA2_TEST_CASES = [
-    PIPE_JINJA2,
+# Export all PipeCompose test cases
+PIPE_COMPOSE_TEST_CASES = [
+    PIPE_COMPOSE,
 ]