+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/pipe_operators/pipe_jinja2/test_data.py
@@ -2,27 +2,27 @@ from typing import ClassVar
 
 from pipelex.core.pipes.pipe_input_blueprint import InputRequirementBlueprint
 from pipelex.libraries.pipelines.builder.pipe.inputs_spec import InputRequirementSpec
-from pipelex.libraries.pipelines.builder.pipe.pipe_jinja2_spec import PipeJinja2Spec, PromptingStyleSpec
-from pipelex.pipe_operators.jinja2.pipe_jinja2_blueprint import PipeJinja2Blueprint
+from pipelex.libraries.pipelines.builder.pipe.pipe_compose_spec import PipeComposeSpec, PromptingStyleSpec
+from pipelex.pipe_operators.compose.pipe_compose_blueprint import PipeComposeBlueprint
 from pipelex.tools.templating.jinja2_template_category import Jinja2TemplateCategory
 from pipelex.tools.templating.templating_models import PromptingStyle, TagStyle, TextFormat
 
 
-class PipeJinja2TestCases:
+class PipeComposeTestCases:
     SIMPLE_JINJA2 = (
         "simple_jinja2",
-        PipeJinja2Spec(
+        PipeComposeSpec(
             the_pipe_code="template_renderer",
             definition="Render a template",
             inputs={"data": InputRequirementSpec(concept="Data")},
             output="RenderedText",
             jinja2="Hello {{ data.name }}!",
         ),
-        PipeJinja2Blueprint(
+        PipeComposeBlueprint(
             definition="Render a template",
             inputs={"data": InputRequirementBlueprint(concept="Data")},
             output="RenderedText",
-            type="PipeJinja2",
+            type="PipeCompose",
             category="PipeOperator",
             jinja2_name=None,
             jinja2="Hello {{ data.name }}!",
@@ -34,7 +34,7 @@ class PipeJinja2TestCases:
 
     JINJA2_WITH_STYLE = (
         "jinja2_with_style",
-        PipeJinja2Spec(
+        PipeComposeSpec(
             the_pipe_code="styled_template",
             definition="Template with prompting style",
             inputs={"input": InputRequirementSpec(concept="Input")},
@@ -47,11 +47,11 @@ class PipeJinja2TestCases:
             template_category=Jinja2TemplateCategory.MARKDOWN,
             extra_context={"version": "1.0"},
         ),
-        PipeJinja2Blueprint(
+        PipeComposeBlueprint(
             definition="Template with prompting style",
             inputs={"input": InputRequirementBlueprint(concept="Input")},
             output="Output",
-            type="PipeJinja2",
+            type="PipeCompose",
             category="PipeOperator",
             jinja2_name="custom_template",
             jinja2=None,
@@ -64,7 +64,7 @@ class PipeJinja2TestCases:
         ),
     )
 
-    TEST_CASES: ClassVar[list[tuple[str, PipeJinja2Spec, PipeJinja2Blueprint]]] = [
+    TEST_CASES: ClassVar[list[tuple[str, PipeComposeSpec, PipeComposeBlueprint]]] = [
         SIMPLE_JINJA2,
         JINJA2_WITH_STYLE,
     ]